+++ cg/cli/upload/genotype.py
@@ -19,8 +19,12 @@ LOG = logging.getLogger(__name__)
 def genotypes(context, re_upload, family_id):
     """Upload genotypes from an analysis to Genotype."""
 
+    if not context.obj.get("analysis_api"):
+        context.obj["analysis_api"] = MipDNAAnalysisAPI(context.obj)
+    analysis_api = context.obj["analysis_api"]
+
     click.echo(click.style("----------------- GENOTYPES -------------------"))
-    analysis_api: MipDNAAnalysisAPI = context.obj["analysis_api"]
+
     if not family_id:
         suggest_cases_to_upload(context)
         context.abort()