+++ cg/constants/gene_panel.py
@@ -1,21 +1,26 @@
 """Gene panel specific constants"""
 
+GENOME_BUILD_37 = "37"
+GENOME_BUILD_38 = "GRCh38"
+
 MASTER_LIST = (
     "BRAIN",
     "Cardiology",
     "CTD",
+    "DIAB",
     "ENDO",
     "EP",
     "IBMFS",
     "IEM",
     "IF",
-    "NEURODEG",
-    "NMD",
     "mcarta",
     "MHT",
     "MIT",
     "MOVE",
     "mtDNA",
+    "NBS-M",
+    "NEURODEG",
+    "NMD",
     "PEDHEP",
     "PID",
     "PIDCAD",