+++ cg/models/compression_data.py
@@ -62,9 +62,7 @@ class CompressionData:
         LOG.info("Check if FASTQ pair exists")
         if not self.file_exists_and_is_accesible(self.fastq_first):
             return False
-        if not self.file_exists_and_is_accesible(self.fastq_second):
-            return False
-        return True
+        return bool(self.file_exists_and_is_accesible(self.fastq_second))
 
     @staticmethod
     def is_absolute(file_path: Path) -> bool: