+++ tests/cli/upload/conftest.py
@@ -3,22 +3,20 @@ import json
 import logging
 from datetime import datetime
 from pathlib import Path
+from tempfile import tempdir
 
 import pytest
-
-from cg.meta.report.api import ReportAPI
-from cg.meta.workflow.mip_dna import MipDNAAnalysisAPI
-from tests.meta.upload.scout.conftest import fixture_mip_load_config
-from tests.mocks.hk_mock import MockHousekeeperAPI
-from tests.mocks.madeline import MockMadelineAPI
-
 from cg.apps.gt import GenotypeAPI
 from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.apps.scout.scoutapi import ScoutAPI
-from cg.meta.upload.scout.scout_load_config import ScoutLoadConfig
+from cg.meta.report.api import ReportAPI
 from cg.meta.upload.scout.scoutapi import UploadScoutAPI
 from cg.meta.workflow.mip import MipAnalysisAPI
+from cg.meta.workflow.mip_dna import MipDNAAnalysisAPI
+from cg.models.scout.scout_load_config import ScoutLoadConfig
 from cg.store import Store, models
+from tests.mocks.hk_mock import MockHousekeeperAPI
+from tests.mocks.madeline import MockMadelineAPI
 
 LOG = logging.getLogger(__name__)
 
@@ -125,27 +123,24 @@ def fixture_base_cli_context(
 ) -> dict:
     """context to use in cli"""
     return {
+        "housekeeper_api": housekeeper_api,
+        "mip-rd-dna": {"root": tempdir},
+        "report_api": MockReportApi(),
         "scout_api": MockScoutApi(),
         "scout_upload_api": upload_scout_api,
-        "housekeeper_api": housekeeper_api,
-        "trailblazer_api": trailblazer_api,
         "status_db": analysis_store,
-        "mip-rd-dna": {"root": "hej"},
+        "trailblazer_api": trailblazer_api,
     }
 
 
-@pytest.fixture(scope="function", name="vogue_context")
-def fixture_vogue_cli_context(vogue_api) -> dict:
-    """context to use in cli"""
+@pytest.fixture(scope="function", name="upload_scout_api")
+def fixture_upload_scout_api(housekeeper_api: MockHousekeeperAPI, mip_load_config: ScoutLoadConfig):
+    """Return a upload scout api"""
+    api = MockScoutUploadApi()
+    api.housekeeper = housekeeper_api
+    api.config = mip_load_config
 
-    return {"vogue_api": vogue_api}
-
-
-@pytest.fixture(scope="function", name="vogue_api")
-def fixture_vogue_api():
-    """Return a MockVogueApi"""
-
-    return MockVogueApi()
+    return api
 
 
 class MockScoutApi(ScoutAPI):
@@ -158,19 +153,19 @@ class MockScoutApi(ScoutAPI):
         LOG.info("Case loaded successfully to Scout")
 
 
-class MockVogueApi:
+class MockReportApi(ReportAPI):
     def __init__(self):
         """docstring for __init__"""
         pass
 
-    def load_reagent_labels(self, days: int):
-        """docstring for upload"""
+    def create_delivery_report(self, *args, **kwargs):
+        """docstring for create_delivery_report"""
 
-    def load_samples(self, days: int):
-        """docstring for upload"""
+        for arg in args:
+            LOG.info("create_delivery_report called with positional %s", arg)
 
-    def load_flowcells(self, days: int):
-        """docstring for upload"""
+        for key, value in kwargs.items():
+            LOG.info("create_delivery_report called with key %s and value %s", key, value)
 
 
 class MockAnalysisApi(MipAnalysisAPI):