+++ tests/meta/orders/test_meta_orders_status.py
@@ -128,12 +128,16 @@ def test_families_to_status(mip_order_to_submit):
     assert family["data_analysis"] == str(Pipeline.MIP_DNA)
     assert family["data_delivery"] == str(DataDelivery.SCOUT)
     assert family["priority"] == "standard"
+    assert family["cohorts"] == {"Other"}
+    assert family["synopsis"] == {"Här kommer det att komma en väldigt lång text med för synopsis."}
     assert set(family["panels"]) == {"IEM"}
     assert len(family["samples"]) == 3
 
     first_sample = family["samples"][0]
+    assert first_sample["age_at_sampling"] == "17.18192"
     assert first_sample["name"] == "sample1"
     assert first_sample["application"] == "WGTPCFC030"
+    assert first_sample["phenotype_terms"] == ["HP:0012747", "HP:0025049"]
     assert first_sample["sex"] == "female"
     assert first_sample["status"] == "affected"
     assert first_sample["mother"] == "sample2"
@@ -145,9 +149,9 @@ def test_families_to_status(mip_order_to_submit):
 
 def test_store_rml(orders_api, base_store, rml_status_data):
     # GIVEN a basic store with no samples and a rml order
-    assert base_store.pools(customer=None).count() == 0
-    assert base_store.families(customer=None).count() == 0
-    assert base_store.samples(customer=None).count() == 0
+    assert base_store.pools().count() == 0
+    assert base_store.families().count() == 0
+    assert base_store.samples().count() == 0
 
     # WHEN storing the order
     new_pools = orders_api.store_rml(
@@ -160,9 +164,9 @@ def test_store_rml(orders_api, base_store, rml_status_data):
     # THEN it should update the database with new pools
     assert len(new_pools) == 2
 
-    assert base_store.pools(customer=None).count() == base_store.families(customer=None).count()
-    assert base_store.samples(customer=None).count() == 4
-    new_pool = base_store.pools(customer=None).first()
+    assert base_store.pools().count() == base_store.families().count()
+    assert base_store.samples().count() == 4
+    new_pool = base_store.pools().first()
 
     assert new_pool == new_pools[1]
 
@@ -174,7 +178,7 @@ def test_store_rml(orders_api, base_store, rml_status_data):
     assert len(new_pool.deliveries) == 1
     assert new_pool.deliveries[0].destination == "caesar"
 
-    new_case = base_store.families(customer=None).first()
+    new_case = base_store.families().first()
     assert new_case.data_analysis == str(Pipeline.FLUFFY)
     assert new_case.data_delivery == str(DataDelivery.NIPT_VIEWER)
 
@@ -186,7 +190,7 @@ def test_store_rml(orders_api, base_store, rml_status_data):
 
 def test_store_rml_bad_apptag(orders_api, base_store, rml_status_data):
     # GIVEN a basic store with no samples and a rml order
-    assert base_store.pools(customer=None).count() == 0
+    assert base_store.pools().count() == 0
 
     for pool in rml_status_data["pools"]:
         pool["application"] = "nonexistingtag"
@@ -227,7 +231,7 @@ def test_store_samples(orders_api, base_store, fastq_status_data):
     assert family_link.family in base_store.families()
     for sample in new_samples:
         assert len(sample.deliveries) == 1
-    assert family_link.family.data_analysis == Pipeline.FASTQ
+    assert family_link.family.data_analysis
     assert family_link.family.data_delivery == DataDelivery.FASTQ
 
 
@@ -247,6 +251,68 @@ def test_store_samples_sex_stored(orders_api, base_store, fastq_status_data):
 
     # THEN the sample sex should be stored
     assert new_samples[0].sex == "male"
+
+
+def test_store_fastq_samples_non_tumour_wgs_to_mip(orders_api, base_store, fastq_status_data):
+    # GIVEN a basic store with no samples and a non-tumour fastq order as wgs
+    assert base_store.samples().count() == 0
+    assert base_store.families().count() == 0
+    base_store.application(fastq_status_data["samples"][0]["application"]).prep_category = "wgs"
+    fastq_status_data["samples"][0]["tumour"] = False
+
+    # WHEN storing the order
+    new_samples = orders_api.store_fastq_samples(
+        customer=fastq_status_data["customer"],
+        order=fastq_status_data["order"],
+        ordered=dt.datetime.now(),
+        ticket=1234348,
+        samples=fastq_status_data["samples"],
+    )
+
+    # THEN the analysis for the case should be MAF
+    assert new_samples[0].links[0].family.data_analysis == Pipeline.MIP_DNA
+
+
+def test_store_fastq_samples_tumour_wgs_to_fastq(orders_api, base_store, fastq_status_data):
+    # GIVEN a basic store with no samples and a tumour fastq order as wgs
+    assert base_store.samples().count() == 0
+    assert base_store.families().count() == 0
+    base_store.application(fastq_status_data["samples"][0]["application"]).prep_category = "wgs"
+    fastq_status_data["samples"][0]["tumour"] = True
+
+    # WHEN storing the order
+    new_samples = orders_api.store_fastq_samples(
+        customer=fastq_status_data["customer"],
+        order=fastq_status_data["order"],
+        ordered=dt.datetime.now(),
+        ticket=1234348,
+        samples=fastq_status_data["samples"],
+    )
+
+    # THEN the analysis for the case should be FASTQ
+    assert new_samples[0].links[0].family.data_analysis == Pipeline.FASTQ
+
+
+def test_store_fastq_samples_non_wgs_as_fastq(orders_api, base_store, fastq_status_data):
+    # GIVEN a basic store with no samples and a fastq order as non wgs
+    assert base_store.samples().count() == 0
+    assert base_store.families().count() == 0
+    non_wgs_prep_category = "wes"
+    assert base_store.applications(category=non_wgs_prep_category)
+    for sample in fastq_status_data["samples"]:
+        sample["application"] = base_store.applications(category=non_wgs_prep_category)[0].tag
+
+    # WHEN storing the order
+    new_samples = orders_api.store_fastq_samples(
+        customer=fastq_status_data["customer"],
+        order=fastq_status_data["order"],
+        ordered=dt.datetime.now(),
+        ticket=1234348,
+        samples=fastq_status_data["samples"],
+    )
+
+    # THEN the analysis for the case should be fastq (none)
+    assert new_samples[0].links[0].family.data_analysis == Pipeline.FASTQ
 
 
 def test_store_samples_bad_apptag(orders_api, base_store, fastq_status_data):
@@ -391,6 +457,11 @@ def test_store_mip(orders_api, base_store, mip_status_data):
     new_link = new_case.links[0]
     assert new_case.data_analysis == str(Pipeline.MIP_DNA)
     assert new_case.data_delivery == str(DataDelivery.SCOUT)
+    assert set(new_case.cohorts) == {"Other"}
+    assert set(new_case.synopsis) == {
+        "H\u00e4r kommer det att komma en v\u00e4ldigt l\u00e5ng text med f\u00f6r synopsis."
+    }
+
     assert new_link.status == "affected"
     assert new_link.mother.name == "sample2"
     assert new_link.father.name == "sample3"
@@ -400,10 +471,6 @@ def test_store_mip(orders_api, base_store, mip_status_data):
     assert new_link.sample.is_tumour
     assert isinstance(new_case.links[1].sample.comment, str)
 
-    assert set(new_link.sample.cohorts) == {"Other"}
-    assert set(new_link.sample.synopsis) == {
-        "H\u00e4r kommer det att komma en v\u00e4ldigt l\u00e5ng text med f\u00f6r synopsis."
-    }
     assert set(new_link.sample.phenotype_terms) == {"HP:0012747", "HP:0025049"}
 
     assert new_link.sample.age_at_sampling == 17.18192